requirements.txt
fastapi
SpeechRecognition
uvicorn
python-dotenv
python-multipart
pyttsx3
google-generativeai
aiofiles